CRoboPeakLidar turns on/off the RPLidar motor

+++ libs/hwdrivers/src/rplidar/src/hal/event.h
@@ -2,26 +2,26 @@
  * Copyright (c) 2014, RoboPeak
  * All rights reserved.
  *
- * Redistribution and use in source and binary forms, with or without
+ * Redistribution and use in source and binary forms, with or without 
  * modification, are permitted provided that the following conditions are met:
  *
- * 1. Redistributions of source code must retain the above copyright notice,
+ * 1. Redistributions of source code must retain the above copyright notice, 
  *    this list of conditions and the following disclaimer.
  *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
+ * 2. Redistributions in binary form must reproduce the above copyright notice, 
+ *    this list of conditions and the following disclaimer in the documentation 
  *    and/or other materials provided with the distribution.
  *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
- * THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
- * PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
- * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
- * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
- * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
- * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
- * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
- * OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" 
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, 
+ * THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR 
+ * PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR 
+ * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, 
+ * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, 
+ * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; 
+ * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, 
+ * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR 
+ * OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, 
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *
  */
@@ -31,7 +31,7 @@
  *
  *  Copyright 2009 - 2014 RoboPeak Team
  *  http://www.robopeak.com
- *
+ * 
  */
 
 
@@ -41,14 +41,14 @@ namespace rp{ namespace hal{
 class Event
 {
 public:
-
+    
     enum
     {
         EVENT_OK = 1,
         EVENT_TIMEOUT = -1,
         EVENT_FAILED = 0,
     };
-
+    
     Event(bool isAutoReset = true, bool isSignal = false)
 #ifdef _WIN32
         : _event(NULL)
@@ -58,7 +58,7 @@ public:
 #endif
     {
 #ifdef _WIN32
-        _event = CreateEvent(NULL, isAutoReset?FALSE:TRUE, isSignal?TRUE:FALSE, NULL);
+        _event = CreateEvent(NULL, isAutoReset?FALSE:TRUE, isSignal?TRUE:FALSE, NULL); 
 #else
         pthread_mutex_init(&_cond_locker, NULL);
         pthread_cond_init(&_cond_var, NULL);
@@ -77,7 +77,7 @@ public:
             SetEvent(_event);
 #else
             pthread_mutex_lock(&_cond_locker);
-
+               
             if ( _is_signalled == false )
             {
                 _is_signalled = true;
@@ -97,8 +97,8 @@ public:
 #endif
         }
     }
-
-    int wait( unsigned long timeout = 0xFFFFFFFF )
+    
+    unsigned long wait( unsigned long timeout = 0xFFFFFFFF )
     {
 #ifdef _WIN32
         switch (WaitForSingleObject(_event, timeout==0xFFFFFFF?INFINITE:(DWORD)timeout))
@@ -112,12 +112,12 @@ public:
         }
         return EVENT_OK;
 #else
-        int ans = EVENT_OK;
+        unsigned long ans = EVENT_OK;
         pthread_mutex_lock( &_cond_locker );
 
         if ( !_is_signalled )
         {
-
+            
                 if (timeout == 0xFFFFFFFF){
                     pthread_cond_wait(&_cond_var,&_cond_locker);
                 }else
@@ -128,7 +128,7 @@ public:
 
                     wait_time.tv_sec = timeout/1000 + now.tv_sec;
                     wait_time.tv_nsec = (timeout%1000)*1000000ULL + now.tv_usec*1000;
-
+                
                     if (wait_time.tv_nsec >= 1000000000)
                     {
                        ++wait_time.tv_sec;
@@ -148,10 +148,10 @@ public:
                         ans = EVENT_FAILED;
                         goto _final;
                     }
-
+       
             }
         }
-
+          
         assert(_is_signalled);
 
         if ( _isAutoReset )
@@ -163,7 +163,7 @@ _final:
 
         return ans;
 #endif
-
+        
     }
 protected:
